- Fixed bad API error when posting. Now all syntaxes correspond to correct API name. - Cleaned up code by moving frequently used functions (such as isInternetAvailable and Load/Save functions) into separate class/file.
